fixing to new osx version

+++ robo3t/install.sh
@@ -6,8 +6,15 @@ source $(dirname $0)/../secrets/dots.sh
 set -e
 
 function move_robo3t_connections_from() {
-  mv ~/.3T/robo-3t/1.2.1/robo3t.json  ~/.3T/robo-3t/1.2.1/robo3t.json.original
-  cp $1/robo3t/robo3t.json ~/.3T/robo-3t/1.2.1/robo3t.json
+  local -r robo3t="~/.3T"
+
+  if [[ -d "${robo3t}" || -L "${robo3t}" ]]; then
+    if [ -f ~/.3T/robo-3t/1.2.1/robo3t.json ]; then
+      mv ~/.3T/robo-3t/1.2.1/robo3t.json  ~/.3T/robo-3t/1.2.1/robo3t.json.original
+    fi
+
+    cp $1/robo3t/robo3t.json ~/.3T/robo-3t/1.2.1/robo3t.json
+  fi
 }
 
 if test ! $(pgrep -f "Robo\ 3T")
